+++ lib/base/game_component.dart
@@ -58,6 +58,9 @@ abstract class GameComponent extends Component
   /// angle is positive, or counterclockwise if the angle is negative.
   double angle = 0;
 
+  bool isFlipVertical = false;
+  bool isFlipHorizontal = false;
+
   /// Param checks if this component is visible on the screen
   bool isVisible = false;
 
@@ -70,7 +73,7 @@ abstract class GameComponent extends Component
     ..style = PaintingStyle.stroke;
 
   TextPaint get debugTextPaint => TextPaint(
-        config: TextPaintConfig(
+        style: TextStyle(
           color: debugColor,
           fontSize: 12,
         ),
@@ -187,15 +190,31 @@ abstract class GameComponent extends Component
     super.renderDebugMode(canvas);
   }
 
-  @override
-  void preRender(Canvas canvas) {
-    if (angle != 0) {
-      canvas.save();
+  void renderTree(Canvas canvas) {
+    canvas.save();
+
+    if (isFlipHorizontal || isFlipVertical || angle != 0) {
       canvas.translate(position.center.dx, position.center.dy);
-      canvas.rotate(angle);
+      if (angle != 0) {
+        canvas.rotate(angle);
+      }
+      if (isFlipHorizontal || isFlipVertical) {
+        canvas.scale(isFlipHorizontal ? -1 : 1, isFlipVertical ? -1 : 1);
+      }
+
       canvas.translate(-position.center.dx, -position.center.dy);
     }
-    super.preRender(canvas);
+
+    render(canvas);
+
+    canvas.restore();
+
+    children.forEach((c) => c.renderTree(canvas));
+
+    // Any debug rendering should be rendered on top of everything
+    if (debugMode) {
+      renderDebugMode(canvas);
+    }
   }
 
   /// Returns true if for each time the defined millisecond interval passes.