+++ lib/map/tile/tile_model.dart
@@ -113,7 +113,10 @@ class TileModel {
   final TileModelSprite? sprite;
   final TileModelAnimation? animation;
   final List<CollisionArea>? collisions;
-  String id = '';
+  final double angle;
+  final bool isFlipVertical;
+  final bool isFlipHorizontal;
+  final String id;
 
   Offset center = Offset.zero;
 
@@ -129,12 +132,15 @@ class TileModel {
     this.sprite,
     this.animation,
     this.collisions,
+    this.angle = 0,
+    this.isFlipVertical = false,
+    this.isFlipHorizontal = false,
+    this.id = '',
   }) {
     center = Offset(
       (x * width) + (width / 2.0),
       (y * height) + (height / 2.0),
     );
-    id = '$x/$y:${DateTime.now().microsecondsSinceEpoch}';
   }
   double get left => (x * width);
   double get right => (x * width) + width;
@@ -158,6 +164,10 @@ class TileModel {
           type: type,
           properties: properties,
         );
+        tile.angle = angle;
+        tile.isFlipHorizontal = isFlipHorizontal;
+        tile.isFlipVertical = isFlipVertical;
+
         tile.gameRef = gameRef;
         tile.id = id;
 
@@ -176,6 +186,9 @@ class TileModel {
           type: type,
           properties: properties,
         );
+        tile.angle = angle;
+        tile.isFlipHorizontal = isFlipHorizontal;
+        tile.isFlipVertical = isFlipVertical;
 
         tile.gameRef = gameRef;
         tile.id = id;
@@ -200,6 +213,9 @@ class TileModel {
           type: type,
           properties: properties,
         );
+        tile.angle = angle;
+        tile.isFlipHorizontal = isFlipHorizontal;
+        tile.isFlipVertical = isFlipVertical;
 
         tile.gameRef = gameRef;
         tile.id = id;
@@ -221,6 +237,10 @@ class TileModel {
           type: type,
           properties: properties,
         );
+        tile.angle = angle;
+        tile.isFlipHorizontal = isFlipHorizontal;
+        tile.isFlipVertical = isFlipVertical;
+
         tile.gameRef = gameRef;
         tile.id = id;