+++ lib/tiled/tiled_world_map.dart
@@ -29,6 +29,13 @@ typedef ObjectBuilder = GameComponent Function(
 class TiledWorldMap {
   static const ORIENTATION_SUPPORTED = 'orthogonal';
   static const ABOVE_TYPE = 'above';
+  static const GIT_ROTATE_180 = 3221225472;
+  static const GIT_ROTATE_90 = 2684354560;
+  static const GIT_ROTATE_270 = 1610612736;
+  static const GIT_FLIP_HORIZONTAL = 2147483648;
+  static const GIT_FLIP_VERTICAL = 1073741824;
+  static const GIT_FLIP_HORIZONTAL_270 = 536870912;
+  static const GIT_FLIP_HORIZONTAL_90 = 3758096384;
   final String path;
   final Size? forceTileSize;
   final ValueChanged<Object>? onError;
@@ -148,6 +155,7 @@ class TiledWorldMap {
   ) {
     _tiles.add(
       TileModel(
+        id: data.id.toString(),
         x: _getX(count, tileLayer.width?.toInt() ?? 0),
         y: _getY(count, tileLayer.width?.toInt() ?? 0),
         offsetX: offsetX,
@@ -159,6 +167,9 @@ class TiledWorldMap {
         sprite: data.sprite,
         properties: data.properties?.isEmpty == true ? null : data.properties,
         type: data.type,
+        angle: data.angle,
+        isFlipVertical: data.isFlipVertical,
+        isFlipHorizontal: data.isFlipHorizontal,
       ),
     );
   }
@@ -211,7 +222,38 @@ class TiledWorldMap {
     return (index / width).floor().toDouble();
   }
 
-  TiledItemTileSet? _getDataTile(int index) {
+  TiledItemTileSet? _getDataTile(int gid) {
+    int index = 0;
+    double angle = 0;
+    bool isFlipX = false;
+    bool isFlipY = false;
+    if (gid > GIT_FLIP_HORIZONTAL_90) {
+      isFlipX = true;
+      angle = 1.5708;
+      index = gid - GIT_FLIP_HORIZONTAL_90;
+    } else if (gid > GIT_ROTATE_180) {
+      angle = 3.14159;
+      index = gid - GIT_ROTATE_180;
+    } else if (gid > GIT_ROTATE_90) {
+      angle = 1.5708;
+      index = gid - GIT_ROTATE_90;
+    } else if (gid > GIT_FLIP_HORIZONTAL) {
+      isFlipX = true;
+      index = gid - GIT_FLIP_HORIZONTAL;
+    } else if (gid > GIT_ROTATE_270) {
+      angle = 4.71239;
+      index = gid - GIT_ROTATE_270;
+    } else if (gid > GIT_FLIP_VERTICAL) {
+      isFlipY = true;
+      index = gid - GIT_FLIP_VERTICAL;
+    } else if (gid > GIT_FLIP_HORIZONTAL_270) {
+      isFlipX = true;
+      angle = 4.71239;
+      index = gid - GIT_FLIP_HORIZONTAL_270;
+    } else {
+      index = gid;
+    }
+
     TileSet? tileSetContain;
     String _pathTileset = '';
     int firsTgId = 0;
@@ -269,11 +311,15 @@ class TiledWorldMap {
       );
 
       return TiledItemTileSet(
+        id: index,
         type: object.type,
         collisions: object.collisions,
         properties: object.properties,
         sprite: sprite,
         animation: animation,
+        angle: angle,
+        isFlipHorizontal: isFlipX,
+        isFlipVertical: isFlipY,
       );
     } else {
       return null;