+++ lib/util/game_color_filter.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/material.dart';
 import 'package:flutter/painting.dart';
 
 class GameColorFilter {